@@ -20,13 +20,11 @@
     "bounce": "rm -rf /node_modules && yarn",
     "build": "webpack -p --mode production",
     "dev": "clear && webpack-dev-server --hot --open --mode development",
-    "eslint": "eslint ./src --fix",
-    "jest": "clear && jest --coverage",
-    "test": "yarn eslint && jest"
+    "lint": "eslint ./src --fix"
   },
   "devDependencies": {
-    "babel-eslint": "^8.2.6",
-    "babel-jest": "^23.4.0",
+    "babel-core": "^6.26.3",
+    "babel-eslint": "^10.0.3",
     "babel-loader": "7.1.5",
     "babel-plugin-transform-class-properties": "^6.24.1",
     "babel-plugin-transform-object-assign": "^6.22.0",
@@ -36,31 +34,30 @@
     "babel-plugin-transform-runtime": "^6.23.0",
     "babel-preset-env": "^1.7.0",
     "babel-preset-react": "^6.24.1",
-    "babelify": "8.0.0",
-    "css-loader": "1.0.0",
-    "eslint": "^5.10.0",
-    "eslint-loader": "^2.1.1",
-    "eslint-plugin-import": "^2.14.0",
-    "eslint-plugin-react": "^7.10.0",
+    "babelify": "10.0.0",
+    "css-loader": "3.2.0",
+    "eslint": "^6.6.0",
+    "eslint-loader": "^3.0.2",
+    "eslint-plugin-import": "^2.18.2",
+    "eslint-plugin-react": "^7.16.0",
     "extract-text-webpack-plugin": "^4.0.0-beta.0",
     "html-webpack-plugin": "3.2.0",
-    "jest": "^23.4.1",
-    "less": "3.7.1",
-    "less-loader": "^4.0.5",
-    "style-loader": "0.21.0",
-    "uglifyjs-webpack-plugin": "1.2.7",
-    "webpack": "^4.29.3",
-    "webpack-bundle-analyzer": "^3.0.3",
-    "webpack-bundle-size-analyzer": "^2.7.0",
-    "webpack-cli": "3.1.0",
-    "webpack-common-shake": "^1.5.3",
-    "webpack-dev-server": "^3.1.14",
+    "less": "3.10.3",
+    "less-loader": "^5.0.0",
+    "style-loader": "1.0.0",
+    "uglifyjs-webpack-plugin": "2.2.0",
+    "webpack": "^4.41.2",
+    "webpack-bundle-analyzer": "^3.6.0",
+    "webpack-bundle-size-analyzer": "^3.1.0",
+    "webpack-cli": "3.3.10",
+    "webpack-common-shake": "^2.1.0",
+    "webpack-dev-server": "^3.9.0",
     "webpack-node-externals": "^1.7.2"
   },
   "dependencies": {
     "granim": "^2.0.0",
-    "react": "^16.8.1",
-    "react-dom": "^16.8.1"
+    "react": "^16.11.0",
+    "react-dom": "^16.11.0"
   },
   "eslintConfig": {
     "extends": "react-app"
